requirements.txt
fastapi
uvicorn
ultralytics
python-multipart
pydantic-settings
Pillow
sqlmodel
websockets